dockerfiles: Use same default image name as DockerHub
To simplify usage of build.sh on CI, use the same image names for local
development as on the Docker hub (i.e. labgrid/client instead of
labgrid-client).

Allow both the prefix and the tag to be overridden using environment
variables.

Use the same prefix in docker-compose.yml, customizable using the same
environment variable.

Adjust the documentation to reflect the new prefix. It is no longer
necessary to build the images locally before running docker-compose so
remove this part of the staging instructions.

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>