github/workflows: Build Docker images on release creation
It is observed that there are no images on Docker hub for Labgrid
release tags, despite listening to tag events.

This may be because this event is not triggered for tags created via the
Github release UI, so add this event too.

Additionally allow the workflow to be triggered manually for testing
purposes.

Signed-off-by: Peter Hoyes <Peter.Hoyes@arm.com>