qemudriver: add support for netdev option in `add_port_forward`
In case there are more `user` network interfaces defined, allow to
specific the netdev. If omitted, an empty string is passed and the
behaviour is like before.

This is useful for testing OpenWrt inside a QEMU instance, since OpenWrt
expects by default a LAN interface on eth0 and a WAN interface (with
uplink) on eth1. Previously the port forward would always be added to
eth1, which doesn't support SSH due to firewall policies. By adding the
netdev to the forward function, the SSHDriver works as expected.

Signed-off-by: Paul Spooren <mail@aparcar.org>